#Zakładka 1 - ID = NULL error, change BIGSERIAL

+++ src/main/resources/db/migration/V1.002__create_author_table.sql
@@ -15,8 +15,8 @@ VALUES (1, 'Jack', 'Hoppkins', 'male', '01-02-2000'),
 
 CREATE TABLE book_author
 (
-    book_id   INT NOT NULL,
-    author_id INT NOT NULL,
+    book_id   BIGSERIAL NOT NULL,
+    author_id BIGSERIAL NOT NULL,
     PRIMARY KEY (book_id, author_id),
     FOREIGN KEY (book_id) REFERENCES book (id),
     FOREIGN KEY (author_id) REFERENCES authors (id)
